[CL-1941] Extract locale switcher
